+++ bin/stats2multiqc.sh
@@ -58,12 +58,21 @@ do
     #ALIGNMENT
     if [ $aligner == "bowtie2" ]; then
 	nb_reads=$(grep "reads;" mapping/${sample}_bowtie2.log | sed 's/ .*//')
-	nb_uniq_reads=$(grep "exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
-	perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	nb_mult_reads=$(grep ">1" mapping/${sample}_bowtie2.log | awk '{print $1}')
-	perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-	perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	if [[ $is_pe == "1" ]]; then
+           nb_uniq_reads=$(grep "concordantly exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
+           perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')                                                                                                   
+           nb_mult_reads=$(grep "concordantly >1" mapping/${sample}_bowtie2.log | awk '{print $1}')
+           perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')                                                                                                   
+           nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))                                                                                                                                                   
+           perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')                                                                                                           
+	else 
+	    nb_uniq_reads=$(grep "exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
+	    perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	    nb_mult_reads=$(grep ">1" mapping/${sample}_bowtie2.log | awk '{print $1}')
+	    perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	    nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
+	    perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	fi
     elif [ $aligner == "bwa-mem" ]; then
 	nb_unmapped=$(grep 'Unmapped' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
 	nb_uniq_reads=$(grep 'Uniquely' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')