+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CAHNGES IN VERSION 1.0.2
 SIGNIFICANT USER-VISIBLE CHANGES
 
   o Update doc
+  o Use paired and single reads in duplicates level calculation
 
 BUG FIXES